+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -6,13 +6,13 @@ import ru.practicum.shareit.user.dto.UserDtoUpdate;
 import java.util.List;
 
 public interface UserService {
-    UserDto getUser(int id);
+    UserDto getUser(Long id);
 
     List<UserDto> getAllUsers();
 
     UserDto create(UserDto userDto);
 
-    UserDto update(int userId, UserDtoUpdate userDto);
+    UserDto update(Long userId, UserDtoUpdate userDto);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 }
